pyproject.toml
[project]
name = "mageai-monitoring"
version = "0.1.0"
description = "GCP Cloud Run service for monitoring Mage.ai pipeline runs and emitting metrics to Cloud Monitoring"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "functions-framework>=3.5.0",
    "flask>=3.0.0",
    "google-cloud-monitoring>=2.16.0",
    "google-cloud-pubsub>=2.18.0",
    "google-cloud-bigquery>=3.13.0",
    "google-cloud-bigquery-storage>=2.22.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "testcontainers>=3.7.0",
    "tenacity>=8.2.0",
    "black>=23.9.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mageai_monitoring"]

[tool.pytest.ini_options]
testpaths = ["tests", "tests_integration"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--cov=.",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
filterwarnings = [
    "ignore::DeprecationWarning:testcontainers.core.waiting_utils",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["N999"]  # Ignore invalid module name (project uses hyphen in name, not underscore)

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

